Split single ufuncs.c file in multiple files
